Comparators
Competitors check if the value is or is not equal to, greater than,
less than or equal too another value.
NOTE == compares whether two things are dual and = assign a value to a
variable.
EQUAL TO (==)
NOT EQUAL TO (!=)
LESS THAN (<)
LESS THAN OR EQUAL TO (<=)
GREATER THAN (>)
GREATER THAN OR EQUAL TO (>=)